test-iot.bbclass: let the set of test cases be extended bitbake-time.
Added a variable to allow the set of test cases in the generated
testplan to be extended dynamically during bitbake time. Simply
append the fully qualified name of test cases to IOTQA_EXTRA_TESTS
and they will be included in the test plan. Test cases are declared
in the form

  "tc1[,tc2[,tc3...]][:image1[,image2[,image3...]]]"

An omitted image (profile) or an explitic '*' will match all
image (profile) names.

For example, here's how to conditionally include an extra test
in image (profile) refkit-image-minimal:

  IOTQA_EXTRA_TESTS_append = " \
      ${@bb.utils.contains('FINGERNAILS', 'long-sharp-sticks', \
           'oeqa.runtime.sadistic.test:refkit-image-minimal', '', d)} \
  "

or to include several tests in several images (profiles):

  IOTQA_EXTRA_TESTS_append = " \
      oeqa.foo.bar1,oeqa.foo.bar2:this-image,that-image \
      oeqa.bar.xyzzy,oeqa.bar.barf:some-other-image \
  "

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>